improv: refactor TestBase
