+++ backend/src/Modules/Auctions/Tests/Test.Auctions.Base/Auctions.Tests.Base.csproj
@@ -11,6 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\..\..\Auctions.DI\Auctions.DI.csproj" />
     <ProjectReference Include="..\..\Auctions.Application\Auctions.Application.csproj" />
     <ProjectReference Include="..\..\Auctions.Domain\Auctions.Domain.csproj" />
   </ItemGroup>
