+++ backend/src/Tests/FunctionalTests/Commands/Auctions/CreateAuction/CreateAuctionCommand_Tests.cs
@@ -1,18 +1,5 @@
-using Adapter.EfCore.ReadModelNotifications;
-using AuctionBids.Domain;
-using AuctionBids.Domain.Repositories;
 using Auctions.Application.Commands.StartAuctionCreateSession;
-using Auctions.Domain.Repositories;
-using Auctions.Tests.Base.Domain.Services.Fakes;
-using Core.Common.Domain.Users;
-using FunctionalTests.Mocks;
-using Microsoft.Extensions.DependencyInjection;
-using MongoDB.Driver;
-using ReadModel.Core.Model;
-using ReadModel.Core.Queries.User.UserAuctions;
-using ReadModel.Core.Queries.User.UserBids;
 using System;
-using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
@@ -20,31 +7,6 @@ using static FunctionalTests.Commands.TestCreateAuctionCommandBuilder;
 
 namespace FunctionalTests.Commands
 {
-    public class CreateAuctionProbe
-    {
-        private readonly TestBase _testBase;
-
-        public CreateAuctionProbe(TestBase testBase)
-        {
-            _testBase = testBase;
-        }
-
-        public bool Check()
-        {
-            using var scope = _testBase.ServiceProvider.GetRequiredService<IServiceScopeFactory>().CreateScope();
-            var auctions = _testBase.ServiceProvider.GetRequiredService<IAuctionRepository>();
-            var auctionBids = _testBase.ServiceProvider.GetRequiredService<IAuctionBidsRepository>();
-            var _readModelNotificationsDbContext = scope.ServiceProvider.GetRequiredService<SagaEventsConfirmationDbContext>();
-            var confirmationsMarkedAsCompleted = _readModelNotificationsDbContext.SagaEventsConfirmations.FirstOrDefault()?.Completed == true;
-            var confirmationEventsProcessed = _readModelNotificationsDbContext.SagaEventsToConfirm.All(e => e.Processed);
-            var createdAuction = _testBase.SendQuery<UserAuctionsQuery, UserAuctionsQueryResult>(new UserAuctionsQuery()).GetAwaiter().GetResult().Auctions.FirstOrDefault();
-            var auctionUnlocked = createdAuction != null && !createdAuction.Locked;
-
-            //if (!confirmationsMarkedAsCompleted) outputHelper.WriteLine("Notifications not marked as completed");
-            return auctionUnlocked && confirmationsMarkedAsCompleted && confirmationEventsProcessed && createdAuction != null;
-        }
-    }
-
     [CollectionDefinition(nameof(CommandTestsCollection), DisableParallelization = true)]
     public class CommandTestsCollection { }
