+++ backend/src/Tests/FunctionalTests/Mocks/UserIdentityServiceMock.cs
@@ -1,18 +1,11 @@
-﻿using Auctions.Application;
-using Common.Application;
-using Moq;
+﻿using Common.Application;
 using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace FunctionalTests.Mocks
 {
     internal class UserIdentityServiceMock : IUserIdentityService
     {
-        public Guid UserId { get; set;  }
+        public Guid UserId { get; set; }
 
         public UserIdentityServiceMock(Guid userId)
         {
@@ -22,17 +15,6 @@ namespace FunctionalTests.Mocks
         public Guid GetSignedInUserIdentity()
         {
             return UserId;
-        }
-    }
-
-    internal static class TestFileStreamAccessor
-    {
-        public static IFileStreamAccessor ForFile(string filePath)
-        {
-            var mock = new Mock<IFileStreamAccessor>();
-            mock.Setup(f => f.GetStream())
-                .Returns(File.OpenRead(filePath));
-            return mock.Object;
         }
     }
 }
